refactor(cli): keep report reader outside artifact core
